Comp bot claw fixes/tests

+++ Team340PowerUp2018/src/org/usfirst/frc/team340/robot/subsystems/Elevator.java
@@ -86,7 +86,7 @@ public class Elevator extends Subsystem {
      * @return the encoder position
      */
     public int getPosition() {
-    	return (int) (talonA.getSelectedSensorPosition(0));
+    	return talonA.getSelectedSensorPosition(0);
     }
     
     /**
